5758 - Revert changes to useGetNationalDataPointValidations
