+++ src/client/pages/Country/SectionWrapper/hooks/useGetNationalDataPointValidations.ts
@@ -1,23 +1,37 @@
 import { useEffect } from 'react'
 
 import { CountryIso } from 'meta/area/countryIso'
+import { SectionNames } from 'meta/assessment/section'
+import { Objects } from 'utils/objects'
 
+import { useCountry } from 'client/store/area/hooks/country'
 import { ValidationsActions } from 'client/store/data/tableData/validations/actions'
 import { useNationalDataPointValidationsFetched } from 'client/store/data/tableData/validations/hooks/nationalDataPoints'
 import { useAppDispatch } from 'client/store/hooks'
+import { useCanEditCycleData } from 'client/store/user/hooks/auth'
+import { useIsDataExportView } from 'client/hooks/dataExport'
 import { useSectionRouteParams } from 'client/hooks/routeParams'
-
-import { useShouldFetchNationalDataPointData } from './useShouldFetchNationalDataPointData'
+import { useIsPrintRoute } from 'client/hooks/routes'
 
 export const useGetNationalDataPointValidations = (): void => {
   const dispatch = useAppDispatch()
   const { assessmentName, countryIso, cycleName, sectionName } = useSectionRouteParams<CountryIso>()
+  const canEdit = useCanEditCycleData()
+  const country = useCountry(countryIso)
+  const isDataExportView = useIsDataExportView()
   const fetched = useNationalDataPointValidationsFetched()
-  const shouldFetch = useShouldFetchNationalDataPointData()
+  const { print } = useIsPrintRoute()
+
+  const hasOriginalDataPoint = Boolean(country?.props?.forestCharacteristics?.useOriginalDataPoint)
+  const isExtentOfForest = sectionName === SectionNames.extentOfForest
+  const isForestCharacteristics = sectionName === SectionNames.forestCharacteristics && hasOriginalDataPoint
+  const isNationalDataPointSection = isExtentOfForest || isForestCharacteristics
+  const shouldFetch =
+    !isDataExportView && !fetched && !print && canEdit && !Objects.isEmpty(sectionName) && isNationalDataPointSection
 
   useEffect(() => {
-    if (!shouldFetch || fetched) return
+    if (!shouldFetch) return
 
     dispatch(ValidationsActions.getNationalDataPointValidations({ assessmentName, cycleName, countryIso, sectionName }))
-  }, [assessmentName, countryIso, cycleName, dispatch, fetched, sectionName, shouldFetch])
+  }, [assessmentName, countryIso, cycleName, dispatch, sectionName, shouldFetch])
 }
